+++ src/main/java/de/jeff_media/InvUnload/UnloadSummary.java
@@ -19,7 +19,8 @@ public class UnloadSummary {
 		unloads = new HashMap<Location,EnumMap<Material,Integer>>();
 	}
 	
-	void unloaded(Location loc, Material mat, int amount) {
+	void protocolUnload(Location loc, Material mat, int amount) {
+		if(amount==0) return;
 		if(!unloads.containsKey(loc)) {
 			unloads.put(loc, new EnumMap<>(Material.class));
 			unloads.get(loc).put(mat, amount);
@@ -37,24 +38,24 @@ public class UnloadSummary {
 		int y = loc.getBlockY();
 		int z = loc.getBlockZ();
 		Material type = loc.getBlock().getType();
-		return String.format(ChatColor.BOLD + "" + ChatColor.DARK_PURPLE + "%s §f@ §aX: §f%d §aY: §f%d §aZ: §f%d", type.name(),x,y,z);
+		return String.format(ChatColor.LIGHT_PURPLE + "§l%s   §a§lX: §f%d §a§lY: §f%d §a§lZ: §f%d", type.name(),x,y,z);
 	}
 	
 	String amount2str(int amount) {
-		return String.format("%6dx  ", amount);
+		return String.format(ChatColor.DARK_PURPLE+"|§7%5dx  ", amount);
 	}
 	
 	void print(PrintRecipient recipient, Player p) {
-		if(unloads.size()>0) printTo(recipient,p," ");
 		
 		for(Entry<Location,EnumMap<Material,Integer>> entry : unloads.entrySet()) {
+			printTo(recipient,p," ");
 			printTo(recipient,p,loc2str(entry.getKey()));
 			EnumMap<Material,Integer> map = entry.getValue();
 			for(Entry<Material,Integer> entry2 : map.entrySet()) {
 				printTo(recipient,p,
-						ChatColor.DARK_GRAY + amount2str(entry2.getValue()) + ChatColor.GRAY + entry2.getKey().name());
+						amount2str(entry2.getValue()) + ChatColor.GOLD + entry2.getKey().name());
 			}
-			printTo(recipient,p," ");
+			//printTo(recipient,p," ");
 		}
 	}